Feature_selection_and_ML/ac_generation/Vaska_ML/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py37h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cryptography=38.0.1=py37h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
docker-pycreds=0.4.0=pypi_0
expat=2.4.9=h6a678d5_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
future=0.18.2=py37_1
giflib=5.2.1=h7b6447c_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
glib=2.69.1=he621ea3_2
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
importlib-metadata=5.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.5.3=py37h06a4308_0
matplotlib-base=3.5.3=py37hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
networkx=2.6.3=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pathlib=1.0.1=py37_2
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pillow=9.2.0=py37hace64e9_1
pip=22.1.2=py37h06a4308_0
ply=3.11=py37_0
promise=2.3=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37_1
python=3.7.15=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.1=cpu_py37he8d8e81_0
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h5eee18b_1
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
schedule=1.1.0=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
sentry-sdk=1.11.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py37h06a4308_0
shortuuid=1.0.11=pypi_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.40.0=h5082296_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchvision=0.13.1=cpu_py37h164cc8f_0
tornado=6.2=py37h5eee18b_0
typing-extensions=4.4.0=py37h06a4308_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.13=py37h06a4308_0
wandb=0.13.6=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
